[wip] Cleanup
