+++ eden/RawTables.h
@@ -1,3 +1,9 @@
+/*
+ *  Everything around raw tables
+ *
+ *
+ */
+
 #ifndef RAWTABLES_H
 #define RAWTABLES_H
 
@@ -20,6 +26,28 @@ typedef void ( *IterationCallback)(
 );
 }
 
+// references to the raw tables
+struct TabEntryRef{
+    long long table;
+    int entry;
+    // could also use the compressed, encoded combination
+};
+typedef long long TabEntryRef_Packed;
+static auto GetEncodedTableEntryId = []( long long global_idx_T_dest_table, long long entry_idx_T_dest ){
+    // pack 1 trillion tables -> 16 million entries into 64bit indexes, upgrade if needed LATER
+
+    const unsigned long long table_id = global_idx_T_dest_table * (1 << 24);
+    const unsigned long long entry_id = entry_idx_T_dest % (1 << 24);
+    const unsigned long long packed_id = table_id | entry_id ;
+
+    return packed_id;
+};
+static auto GetDecodedTableEntryId = [  ]( unsigned long long packed_id ){
+    long long global_idx_T_dest_table = packed_id >> 24;
+    int entry_idx_T_dest = packed_id % (1 << 24);
+    TabEntryRef ret = { global_idx_T_dest_table, entry_idx_T_dest };
+    return ret;
+};
 
 extern "C" {
 // initial states, internal constants, connectivity matrices, iteration function pointers and everything
@@ -93,7 +121,6 @@ struct RawTables{
             }
         }
         consecutive_kernels.push_back(cic);
-//        printf("create_consecutive_kernels_vector : reduced %lld callbacks to %lld consecutive kernels\n", (long long)callbacks.size(), (long long)consecutive_kernels.size());
     }
 };
 }
