+++ eden/StringHelpers.h
